# swagger/swagger.yml
swagger: '2.0'
info:
  title: Calendar
  version: '1.0'
  description: Simple event calendar server
  license:
    name: Open license
host: 'localhost:5000'
tags:
  - name: auth
    description: Basic auth operations
  - name: user
    description: Basic user operations
  - name: events
    description: Basic events operations 
  - name: event
    description: Basic event operations
schemes:
  - http
  - https
paths:
  /login:
    post:
      tags:
       - auth
      summary: Logs user into the system
      description: ''
      consumes:
        - application/json
      parameters:
        - in: body
          name: body
          description: Auth form data as json
          required: true
          schema:
            $ref: '#/definitions/Auth'
      responses:
        '200':
          description: Successfully saved"

  /logout:
    get:
      tags:
        - auth
      summary: Logs out current logged in user session
      description: ''
      operationId: logoutUser
      produces:
        - text/plain
      parameters: []
      responses:
        default:
          description: successful loged out
  /api/user:
    put:
      tags: 
       - user
      summary: Update user's timezone
      description: 'This operation can be done only for loged in users'
      consumes:
        - application/json
      parameters:
        - in: body
          name: body
          description: User's timezone data
          required: true
          schema:
            $ref: '#/definitions/User'
      responses:
       '401':
          description: Unathorized access
       '200':
          description: Successfully saved
  /api/events:
    get:
      tags: 
       - events
      summary: Get events
      description: 'This operation can be done only for loged in users'
      consumes:
        - application/json
      parameters:
        - name: title
          in: query
          description: The title of the event
          required: false
          type: string
          default: birthday
        - name: timezone
          in: query
          description: The timezone of events
          required: false
          type: string
          default: Europe/Riga
        - name: dateFrom
          in: query
          description: Events must be after or on this date
          required: false
          type: string
          format: date
          default: '2021-09-01'
        - name: dateTo
          in: query
          description: Events must be before this date
          required: false
          type: string
          format: date
          default: '2021-09-01'
        - name: timeFrom
          in: query
          description: Events must be after or on this time
          required: false
          type: string
          format: time
          default: 08:00
        - name: timeTo
          in: query
          description: Events must be before this time
          required: false
          type: string
          format: time
          default: 10:00
      responses:
       '401':
          description: Unathorized access
       '200':
          description: Successful operation
          schema:
            type: array
            items:
              $ref: '#/definitions/Event'
  /api/event:
    post:
      tags: 
       - event
      summary: Create event
      description: 'This operation can be done only for loged in users'
      consumes:
        - application/json
      parameters:
        - in: body
          name: body
          description: Created event object
          required: true
          schema:
            $ref: '#/definitions/Event'
      responses:
       '401':
          description: Unathorized access
       '201':
          description: Successfully saved
    put:
      tags: 
       - event
      summary: Update event
      description: 'This operation can be done only for loged in users'
      consumes:
        - application/json
      parameters:
        - in: body
          name: body
          description: Updated event object
          required: true
          schema:
            $ref: '#/definitions/Event'
      responses:
       '401':
          description: Unathorized access
       '201':
          description: Successfully saved
  /api/event/{id}:
    get:
      tags: 
       - event
      summary: Get event by id
      description: 'This operation can be done only for loged in users'
      parameters:
        - name: id
          in: path
          description: 'Event ID'
          required: true
          type: integer
          default: 1
      responses:
       '401':
          description: Unathorized access
       '200':
          description: Successful operation
          schema:
            $ref: '#/definitions/Event'
            
definitions:
  Auth:
    type: object
    properties:
      Username:
        type: string
      Password:
        type: string 
  User:
    type: object
    properties:
      login:
        type: string
      timezone:
        type: string       
        
  Event:
    type: object
    properties: 
      id: 
        type: string
      title:
        type: string
      description:
        type: string
      time:
        type: string
      timezone:
        type: string
      duration:
        type: integer
        format: int32
      notes:
        type: array
        items: 
          type: string